Components of Campanas separated for organization purposes
