+++ src/Data/TestData.jsx
@@ -1,0 +1,140 @@
+export const tareasIniciales = [
+  {
+    id: "1",
+    peticion: "Publicación Calentador solar Raysol en Super precio",
+    medio_origen: "Correo",
+    fecha_limite: "06/04/2023",
+    prioridad: "1-URGENTE",
+    servidor_ubicacion: "",
+    actividad: "Web",
+    uso: "Actualizacion de Datos",
+    // frecuencia: "",
+    // duracion: "",
+    estado: "Correcciones",
+    fecha_entrega: "",
+    info_cliente: "Informacion del cliente",
+    observaciones: "",
+    correcciones: "",
+    error: "",
+    tipo_error: "",
+    empresa: "Gaytan",
+    levanta_ticket: "Maria",
+    cliente_solicita: "Roberto G.",
+    encargado: "Maria",
+    apoyo: "Mario",
+  },
+  {
+    id: "200",
+    peticion: "informacion sobre le ticket 2",
+    medio_origen: "Correo",
+    fecha_limite: "06/04/2023",
+    prioridad: "1-URGENTE",
+    servidor_ubicacion: "",
+    actividad: "Web",
+    uso: "Actualizacion de Datos",
+    // frecuencia: "",
+    // duracion: "",
+    estado: "Correcciones",
+    fecha_entrega: "",
+    info_cliente: "Informacion del cliente",
+    observaciones: "",
+    correcciones: "",
+    error: "",
+    tipo_error: "",
+    empresa: "Gaytan",
+    levanta_ticket: "Maria",
+    cliente_solicita: "Roberto G.",
+    encargado: "Maria",
+    apoyo: "Mario",
+  },
+  {
+    id: "32352",
+    peticion: "Ticket 3: realizar las nuevas modificaciones antes del 24",
+    medio_origen: "Correo",
+    fecha_limite: "06/04/2023",
+    prioridad: "1-URGENTE",
+    servidor_ubicacion: "",
+    actividad: "Web",
+    uso: "Actualizacion de Datos",
+    // frecuencia: "",
+    // duracion: "",
+    estado: "Correcciones",
+    fecha_entrega: "",
+    info_cliente: "Informacion del cliente",
+    observaciones: "",
+    correcciones: "",
+    error: "",
+    tipo_error: "",
+    empresa: "Gaytan",
+    levanta_ticket: "Maria",
+    cliente_solicita: "Roberto G.",
+    encargado: "Maria",
+    apoyo: "Mario",
+  },
+];
+
+export const initialUsers = [
+  {
+    id: "1",
+    username: "Mario",
+    first_name: "Mario Hugo",
+    last_name: "Mendez Villa",
+    group: "creativo",
+    is_active: "true",
+    last_login: "28-02-2023 13:02",
+  },
+  {
+    id: "2",
+    username: "Maria",
+    first_name: "Maria Agata",
+    last_name: "Rodriguez Gomez",
+    group: "creativo",
+    is_active: "true",
+    last_login: "28-04-2023 13:45",
+  },
+  {
+    id: "3",
+    username: "Roberto",
+    first_name: "Roberto",
+    last_name: "Rodriguez Gomez",
+    group: "creativo",
+    is_active: "true",
+    last_login: "28-04-2023 13:45",
+  },
+  {
+    id: "4",
+    username: "Emily",
+    first_name: "Emily Helena",
+    last_name: "Rodriguez Gomez",
+    group: "creativo",
+    is_active: "true",
+    last_login: "28-04-2023 13:45",
+  },
+  {
+    id: "5",
+    username: "Emily",
+    first_name: "Emily Helena",
+    last_name: "Rodriguez Gomez",
+    group: "creativo",
+    is_active: "true",
+    last_login: "28-04-2023 13:45",
+  },
+  {
+    id: "6",
+    username: "Emily",
+    first_name: "Emily Helena",
+    last_name: "Rodriguez Gomez",
+    group: "creativo",
+    is_active: "true",
+    last_login: "28-04-2023 13:45",
+  },
+  {
+    id: "7",
+    username: "Emily",
+    first_name: "Emily Helena",
+    last_name: "Rodriguez Gomez",
+    group: "creativo",
+    is_active: "true",
+    last_login: "28-04-2023 13:45",
+  },
+];
