+++ src/components/LeftDrawer/Side.jsx
@@ -11,7 +11,7 @@ export default function Side({ setIsOpen }) {
           xmlns="http://www.w3.org/2000/svg"
           viewBox="0 0 24 24"
           fill="currentColor"
-          class="h-auto w-auto"
+          className="h-auto w-auto"
         >
           <path d="M3,6H21V8H3V6M3,11H21V13H3V11M3,16H21V18H3V16Z" />
         </svg>
