+++ src/pages/Creativo/Campanas.jsx
@@ -1,147 +1,8 @@
 import React, { useState } from "react";
 import { DragDropContext, Droppable, Draggable } from "@hello-pangea/dnd";
-import AccountCircleIcon from "@mui/icons-material/AccountCircle";
-
-const tareasIniciales = [
-  {
-    id: "1",
-    peticion: "Publicación Calentador solar Raysol en Super precio",
-    medio_origen: "Correo",
-    fecha_limite: "06/04/2023",
-    prioridad: "1-URGENTE",
-    servidor_ubicacion: "",
-    actividad: "Web",
-    uso: "Actualizacion de Datos",
-    // frecuencia: "",
-    // duracion: "",
-    estado: "Correcciones",
-    fecha_entrega: "",
-    info_cliente: "Informacion del cliente",
-    observaciones: "",
-    correcciones: "",
-    error: "",
-    tipo_error: "",
-    empresa: "Gaytan",
-    levanta_ticket: "Maria",
-    cliente_solicita: "Roberto G.",
-    encargado: "Maria",
-    apoyo: "Mario",
-  },
-  {
-    id: "200",
-    peticion: "informacion sobre le ticket 2",
-    medio_origen: "Correo",
-    fecha_limite: "06/04/2023",
-    prioridad: "1-URGENTE",
-    servidor_ubicacion: "",
-    actividad: "Web",
-    uso: "Actualizacion de Datos",
-    // frecuencia: "",
-    // duracion: "",
-    estado: "Correcciones",
-    fecha_entrega: "",
-    info_cliente: "Informacion del cliente",
-    observaciones: "",
-    correcciones: "",
-    error: "",
-    tipo_error: "",
-    empresa: "Gaytan",
-    levanta_ticket: "Maria",
-    cliente_solicita: "Roberto G.",
-    encargado: "Maria",
-    apoyo: "Mario",
-  },
-  {
-    id: "32352",
-    peticion: "Ticket 3: realizar las nuevas modificaciones antes del 24",
-    medio_origen: "Correo",
-    fecha_limite: "06/04/2023",
-    prioridad: "1-URGENTE",
-    servidor_ubicacion: "",
-    actividad: "Web",
-    uso: "Actualizacion de Datos",
-    // frecuencia: "",
-    // duracion: "",
-    estado: "Correcciones",
-    fecha_entrega: "",
-    info_cliente: "Informacion del cliente",
-    observaciones: "",
-    correcciones: "",
-    error: "",
-    tipo_error: "",
-    empresa: "Gaytan",
-    levanta_ticket: "Maria",
-    cliente_solicita: "Roberto G.",
-    encargado: "Maria",
-    apoyo: "Mario",
-  },
-];
-
-const initialUsers = [
-  {
-    id: "1",
-    username: "Mario",
-    first_name: "Mario Hugo",
-    last_name: "Mendez Villa",
-    group: "creativo",
-    is_active: "true",
-    last_login: "28-02-2023 13:02",
-  },
-  {
-    id: "2",
-    username: "Maria",
-    first_name: "Maria Agata",
-    last_name: "Rodriguez Gomez",
-    group: "creativo",
-    is_active: "true",
-    last_login: "28-04-2023 13:45",
-  },
-  {
-    id: "3",
-    username: "Roberto",
-    first_name: "Roberto",
-    last_name: "Rodriguez Gomez",
-    group: "creativo",
-    is_active: "true",
-    last_login: "28-04-2023 13:45",
-  },
-  {
-    id: "4",
-    username: "Emily",
-    first_name: "Emily Helena",
-    last_name: "Rodriguez Gomez",
-    group: "creativo",
-    is_active: "true",
-    last_login: "28-04-2023 13:45",
-  },
-  {
-    id: "5",
-    username: "Emily",
-    first_name: "Emily Helena",
-    last_name: "Rodriguez Gomez",
-    group: "creativo",
-    is_active: "true",
-    last_login: "28-04-2023 13:45",
-  },
-  {
-    id: "6",
-    username: "Emily",
-    first_name: "Emily Helena",
-    last_name: "Rodriguez Gomez",
-    group: "creativo",
-    is_active: "true",
-    last_login: "28-04-2023 13:45",
-  },
-  {
-    id: "7",
-    username: "Emily",
-    first_name: "Emily Helena",
-    last_name: "Rodriguez Gomez",
-    group: "creativo",
-    is_active: "true",
-    last_login: "28-04-2023 13:45",
-  },
-];
+import DraggableTicket from "../../components/DnD/DraggableTicket";
+import Usercard from "../../components/Card/UserCard";
+import { tareasIniciales, initialUsers } from "../../Data/TestData";
 
 const reorder = (list, startIndex, endIndex) => {
   const result = [...list];
@@ -161,6 +22,10 @@ const deleteCard = (list, startIndex) => {
 const Campanas = () => {
   const [tickets, setTickets] = useState(tareasIniciales);
   const [users, setusers] = useState(initialUsers);
+
+  const handleClickTicket = () => {
+    console.log("UwU")
+  }
 
   const handleDragEnd = (result) => {
     const { source, destination } = result;
@@ -191,7 +56,7 @@ const Campanas = () => {
   return (
     <DragDropContext onDragEnd={handleDragEnd}>
       <div className="h-full p-1 bg-slate-100 flex">
-        <div className="h-full w-1/4 m-1 p-2 bg-primary flex-col rounded-md">
+        <div className="h-full w-1/4 m-1 p-2 bg-teal-500 flex-col rounded-md bg-opacity-20">
           <h2 className="m-2 text-center font-bold">Tickets</h2>
           <div>
             <Droppable droppableId="tickets">
@@ -202,7 +67,7 @@ const Campanas = () => {
                   className="flex flex-col"
                 >
                   {tickets.map((ticket, index) => (
-                    <DraggableTicket key={ticket.id} ticket={ticket} index={index} />
+                    <DraggableTicket key={ticket.id} onClick={handleClickTicket} ticket={ticket} index={index} />
                   ))}
                   {droppableProvided.placeholder}
                 </ul>
@@ -212,7 +77,7 @@ const Campanas = () => {
         </div>
 
         <div className="h-full w-3/4 m-1 rounded-md flex flex-col">
-          <div className="h-1/8 w-full rounded-md bg-secondary">
+          <div className="h-1/8 w-full rounded-md bg-teal-500 bg-opacity-20">
             <h2 className="mt-2 pt-2 text-center font-bold">Creativos</h2>
             <div className="px-2 overflow-auto">
               <Droppable droppableId="users">
@@ -233,7 +98,7 @@ const Campanas = () => {
               </Droppable>
             </div>
           </div>
-          <div className="w-full mt-2 p-2 rounded-md bg-secondary flex grow justify-center">
+          <div className="w-full mt-2 p-2 rounded-md bg-teal-800 bg-opacity-20 flex grow justify-center">
             <p className="m-2 text-center self-center font-light">
               Selecciona un Ticket o Usuario
             </p>
@@ -242,106 +107,6 @@ const Campanas = () => {
         </div>
       </div>
     </DragDropContext>
-  );
-};
-
-const DraggableTicket = ({ ticket, index }) => {
-  return (
-    <Draggable draggableId={ticket.id} index={index}>
-      {(draggableProvided) => (
-        <div
-          className={`mx-2 mb-4 p-2 h-fit bg-white rounded-md hover:scale-95 flex flex-col`}
-          ref={draggableProvided.innerRef}
-          {...draggableProvided.draggableProps}
-          {...draggableProvided.dragHandleProps}
-        >
-          <div className="bg-slate-000 flex pb-2">
-            <p className="px-0.5 bg-teal-200 rounded-lg">{ticket.id}</p>
-            <p className="ml-4">{ticket.empresa}</p>
-          </div>
-          <div className="bg-slate-000 flex pb-1">
-            <StateTag state='urgente'/>
-            <ActivityTag activity='Edición Fotografía'/>
-          </div>
-          <div className="bg-slate-000 flex text-sm h-auto">
-            {ticket.peticion}
-          </div>
-          {/* <p>ID: {ticket.id}</p>
-          <p>{ticket.peticion}</p> */}
-        </div>
-      )}
-    </Draggable>
-  );
-};
-
-const Usercard = ({ user, index }) => {
-  return (
-    // <Draggable draggableId={user.id} index={index}>
-    //   {(draggableProvided) => (
-    <div
-      className="m-3 p-1 flex bg-white rounded-md hover:scale-95 w-40"
-      // ref={draggableProvided.innerRef}
-      // {...draggableProvided.draggableProps}
-      // {...draggableProvided.dragHandleProps}
-    >
-      {/* <AccountCircleIcon className="" /> */}
-      <div className="bg-primary w-12 h-12 m-2 rounded-full"></div>
-      <div className="m-2">
-        <p>{user.first_name}</p>
-      </div>
-    </div>
-    //   )}
-    // </Draggable>
-  );
-};
-
-const StateTag = ({ state }) => {
-  const [bgColor, setBgColor] = useState(() => {
-    switch (state) {
-      case 'urgente':
-        return 'bg-red-600';
-      case 'importante':
-        return 'bg-yellow-500';
-      case 'normal':
-        return 'bg-green-400';
-      case 'programado':
-        return 'bg-blue-500';
-      default:
-        return 'bg-gray-400';
-    }
-  });
-
-  return (
-    <div
-      className={`rounded-md text-white font-bold text-xs px-1 h-5 ${bgColor}`}
-    >
-      {state}
-    </div>
-  );
-};
-
-const ActivityTag = ({ activity }) => {
-  const [bgColor, setBgColor] = useState(() => {
-    switch (activity) {
-      case 'Web':
-        return 'bg-red-400';
-      case 'Evento':
-        return 'bg-yellow-400';
-      case '3D':
-        return 'bg-green-300';
-      case 'Edición Fotografía':
-        return 'bg-blue-300';
-      default:
-        return 'bg-gray-400';
-    }
-  });
-
-  return (
-    <div
-      className={`rounded-md text-white font-bold text-xs px-1 h-5 ml-2 ${bgColor}`}
-    >
-      {activity}
-    </div>
   );
 };
